Add a rake task for seeding the database with superfun data. (#93)
* WIP

* Add a rake task for seeding the database with a bunch of stuff

This is to help make the app a bit more attractive in demos. I also
threw in a couple of field additions to hazards, which I realize I
should have done as a separate commit but had lost track of which branch
I was in when making those changes. They're really small, so I think
this is acceptable, just this one time.

* seed the random number generator for consistent results

* add README information about configuring Twilio